Podłaczenie respa przeciwnika po zmianie etapu, podstawowa walka
Po przegranej cofanie do etapu 1 oraz usunięcie wszystkich zebranych grzybów

+++ Assets/Scripts/Enemy/Enemy.cs
@@ -7,7 +7,7 @@ public class Enemy : MonoBehaviour
     public int ID;
     public string Name;
     public int HP;
-    public double Damage;
+    public int Damage;
     public string SpriteName;
 }
